refactor another way
