+++ yt/yt/library/query/engine/public.h
@@ -4,11 +4,4 @@
 
 namespace NYT::NQueryClient {
 
-////////////////////////////////////////////////////////////////////////////////
-
-DECLARE_REFCOUNTED_STRUCT(TCodegenCacheConfig)
-DECLARE_REFCOUNTED_STRUCT(TCodegenCacheDynamicConfig)
-
-////////////////////////////////////////////////////////////////////////////////
-
 } // namespace NYT::NWebAssembly
