+++ yt/yt/library/query/engine/query_engine_config.cpp
@@ -1,4 +1,6 @@
-#include "query_engine_config.h"
+#include "cg_cache.h"
+
+#include <yt/yt/library/query/engine_api/query_engine_config.h>
 
 #include <yt/yt/library/query/base/private.h>
 
@@ -8,46 +10,6 @@ namespace NYT::NQueryClient {
 
 using namespace NYson;
 using namespace NYTree;
-
-////////////////////////////////////////////////////////////////////////////////
-
-#ifdef YT_VERBOSE_CHANGING_QUERY_ENGINE_CONFIG
-
-constinit const auto Logger = QueryClientLogger;
-
-#endif
-
-////////////////////////////////////////////////////////////////////////////////
-
-void TQueryEngineConfig::Register(TRegistrar registrar)
-{
-    registrar.Parameter("codegen_cache", &TThis::CodegenCache)
-        .DefaultNew();
-}
-
-void TQueryEngineDynamicConfig::Register(TRegistrar registrar)
-{
-    registrar.Parameter("codegen_cache", &TThis::CodegenCache)
-        .DefaultNew();
-
-    registrar.Parameter("statistics_aggregation", &TThis::StatisticsAggregation)
-        .Optional();
-
-    registrar.Parameter("use_order_by_in_join_subqueries", &TThis::UseOrderByInJoinSubqueries)
-        .Optional();
-
-    registrar.Parameter("expression_builder_version", &TThis::ExpressionBuilderVersion)
-        .Optional();
-
-    registrar.Parameter("codegen_optimization_level", &TThis::OptimizationLevel)
-        .Optional();
-
-    registrar.Parameter("rewrite_cardinality_into_hyper_log_log_with_precision", &TThis::RewriteCardinalityIntoHyperLogLogWithPrecision)
-        .Optional();
-
-    registrar.Parameter("allow_join_with_async_last_committed_timestamp_if_require_sync_replica_is_false", &TThis::AllowJoinWithAsyncLastCommittedTimestampIfRequireSyncReplicaIsFalse)
-        .Optional();
-}
 
 ////////////////////////////////////////////////////////////////////////////////
