+++ yt/yt/library/query/engine_api/public.h
@@ -32,4 +32,9 @@ const TConstConstraintExtractorMapPtr GetBuiltinConstraintExtractors();
 
 ////////////////////////////////////////////////////////////////////////////////
 
+DECLARE_REFCOUNTED_STRUCT(TCodegenCacheConfig)
+DECLARE_REFCOUNTED_STRUCT(TCodegenCacheDynamicConfig)
+
+////////////////////////////////////////////////////////////////////////////////
+
 } // namespace NYT::NQueryClient
