+++ yt/yt/library/query/engine_api/row_level_security.cpp
@@ -1,0 +1,9 @@
+#include "row_level_security.h"
+
+namespace NYT::NQueryClient {
+
+Y_WEAK std::pair<ICGInstanceHolderPtr, NTableClient::TTableSchemaPtr> MakeRlsCGInstance(const NTableClient::TTableSchemaPtr& /*schema*/, const std::string& /*predicate*/) {
+    YT_ABORT();
+}
+
+}
