--- yt/yt/library/query/row_level_security_api/ya.make
@@ -1,15 +1,0 @@
-LIBRARY()
-
-INCLUDE(${ARCADIA_ROOT}/yt/ya_cpp.make.inc)
-
-SRCS(
-    row_level_security.cpp
-    proto/row_level_security.proto
-)
-
-PEERDIR(
-    yt/yt/core
-    yt/yt/client
-)
-
-END()
