+++ yt/yt/library/query/unittests/evaluate/test_evaluate.cpp
@@ -2,8 +2,7 @@
 
 #include <yt/yt/library/query/base/query_helpers.h>
 
-#include <yt/yt/library/query/engine/query_engine_config.h>
-
+#include <yt/yt/library/query/engine_api/query_engine_config.h>
 #include <yt/yt/library/query/engine_api/coordinator.h>
 #include <yt/yt/library/query/engine_api/evaluator.h>
