+++ yt/yt/server/controller_agent/controllers/table.h
@@ -13,14 +13,13 @@
 #include <yt/yt/ytlib/cypress_client/public.h>
 
 #include <yt/yt/ytlib/table_client/config.h>
+#include <yt/yt/ytlib/table_client/row_level_security.h>
 
 #include <yt/yt/ytlib/scheduler/cluster_name.h>
 #include <yt/yt/ytlib/scheduler/input_query.h>
 
 #include <yt/yt/client/table_client/column_rename_descriptor.h>
 #include <yt/yt/client/table_client/table_upload_options.h>
-
-#include <yt/yt/library/query/row_level_security_api/row_level_security.h>
 
 #include <yt/yt_proto/yt/client/chunk_client/proto/data_statistics.pb.h>
