+++ yt/yt/ytlib/api/native/client_dynamic_tables_impl.cpp
@@ -106,7 +106,7 @@
 #include <yt/yt/library/query/base/functions.h>
 #include <yt/yt/library/query/base/query_preparer.h>
 
-#include <yt/yt/library/query/engine/query_engine_config.h>
+#include <yt/yt/library/query/engine_api/query_engine_config.h>
 
 #include <yt/yt/library/query/engine_api/column_evaluator.h>
 #include <yt/yt/library/query/engine_api/new_range_inferrer.h>
